+++ Improvar_Cloth/Models/T_TXNDTL.cs
@@ -85,73 +85,73 @@ namespace Improvar.Models
         [StringLength(2)]
         public string JOBCD { get; set; }
 
-        public decimal? NOS { get; set; }
+         public double? NOS { get; set; }
 
-        public decimal? QNTY { get; set; }
+         public double? QNTY { get; set; }
 
-        public decimal? BLQNTY { get; set; }
+         public double? BLQNTY { get; set; }
 
-        public decimal? RATE { get; set; }
+         public double? RATE { get; set; }
 
-        public decimal? AMT { get; set; }
+         public double? AMT { get; set; }
 
-        public decimal? FLAGMTR { get; set; }
+         public double? FLAGMTR { get; set; }
 
-        public decimal? TOTDISCAMT { get; set; }
+         public double? TOTDISCAMT { get; set; }
 
-        public decimal? TXBLVAL { get; set; }
+         public double? TXBLVAL { get; set; }
 
-        public decimal? IGSTAMT { get; set; }
+         public double? IGSTAMT { get; set; }
 
-        public decimal? IGSTPER { get; set; }
+         public double? IGSTPER { get; set; }
 
-        public decimal? CGSTAMT { get; set; }
+         public double? CGSTAMT { get; set; }
 
-        public decimal? CGSTPER { get; set; }
+         public double? CGSTPER { get; set; }
 
-        public decimal? SGSTAMT { get; set; }
+         public double? SGSTAMT { get; set; }
 
-        public decimal? SGSTPER { get; set; }
+         public double? SGSTPER { get; set; }
 
-        public decimal? CESSAMT { get; set; }
+         public double? CESSAMT { get; set; }
 
-        public decimal? CESSPER { get; set; }
+         public double? CESSPER { get; set; }
 
-        public decimal? DUTYAMT { get; set; }
+         public double? DUTYAMT { get; set; }
 
-        public decimal? DUTYPER { get; set; }
+         public double? DUTYPER { get; set; }
 
-        public decimal? NETAMT { get; set; }
+         public double? NETAMT { get; set; }
 
-        public decimal? OTHRAMT { get; set; }
+         public double? OTHRAMT { get; set; }
 
         [StringLength(16)]
         public string AGDOCNO { get; set; }
 
         public DateTime? AGDOCDT { get; set; }
 
-        public decimal? SHORTQNTY { get; set; }
+         public double? SHORTQNTY { get; set; }
 
-        public decimal? DISCRATE { get; set; }
+         public double? DISCRATE { get; set; }
 
         [StringLength(1)]
         public string DISCTYPE { get; set; }
 
-        public decimal? DISCAMT { get; set; }
+         public double? DISCAMT { get; set; }
 
-        public decimal? SCMDISCRATE { get; set; }
+         public double? SCMDISCRATE { get; set; }
 
         [StringLength(1)]
         public string SCMDISCTYPE { get; set; }
 
-        public decimal? SCMDISCAMT { get; set; }
+         public double? SCMDISCAMT { get; set; }
 
-        public decimal? TDDISCRATE { get; set; }
+         public double? TDDISCRATE { get; set; }
 
         [StringLength(1)]
         public string TDDISCTYPE { get; set; }
 
-        public decimal? TDDISCAMT { get; set; }
+         public double? TDDISCAMT { get; set; }
 
         [StringLength(4)]
         public string PRCCD { get; set; }
